removed neutronics features
