+++ paramak/utils.py
@@ -1,5 +1,8 @@
 
 import math
+import os
+import shutil
+import subprocess
 from collections.abc import Iterable
 from hashlib import blake2b
 from os import fdopen, remove
@@ -13,8 +16,346 @@ import numpy as np
 import plotly.graph_objects as go
 from cadquery import importers
 from OCP.GCPnts import GCPnts_QuasiUniformDeflection
+from remove_dagmc_tags import remove_tags
 
 import paramak
+
+
+def trelis_command_to_create_dagmc_h5m(
+        faceting_tolerance: float,
+        merge_tolerance: float,
+        material_key_name: Optional[str] = 'material_tag',
+        geometry_key_name: Optional[str] = 'stp_filename',
+        batch: Optional[bool] = True,
+        h5m_filename: str = 'dagmc_not_watertight.h5m',
+        manifest_filename: str = 'manifest.json',
+        cubit_filename: str = 'dagmc.cub',
+        trelis_filename: str = 'dagmc.trelis',
+        geometry_details_filename: str = 'geometry_details.json',
+        surface_reflectivity_name: str = 'reflective',
+) -> List[str]:
+    """Runs the Trelis executable command with the
+    make_faceteted_neutronics_model.py script which produces a non water tight
+    DAGMC h5m file.
+
+    Arguments:
+        faceting_tolerance: the tolerance to use when faceting surfaces.
+        merge_tolerance: the tolerance to use when merging surfaces.
+        material_key_name: the dictionary key containing the str or int to use
+            as the material identifier.
+        geometry_key_name: the dictionary key containing the str to uses as the
+            CAD file identifier.
+        batch: Run the Trelis command in batch model with no GUI (True) or with
+            the GUI enabled (False).
+        h5m_filename: the filename of the DAGMC h5m file produced. This is not
+            water tight at this stage.
+        manifest_filename: The filename of the json file containing a list of
+            material_keys and geometry_keys.
+        cubit_filename: The output filename of the file. If None then no cubit
+            file will be exported.
+        trelis_filename: The output filename of the file. If None then no
+            trelis file will be exported.
+        geometry_details_filename: The output filename of the JSON file
+            containing details of the DAGMC geometry. This includes the
+            resulting volume numbers of the input CAD files, which can be
+            useful for specifying tallies. If None then no JSON fie will be
+            exported.
+        surface_reflectivity_name: The tag to assign to the reflective boundary
+            in the resulting DAGMC geometry Shift requires "spec.reflect" and
+            MCNP requires "boundary:Reflecting".
+
+    Returns:
+        The filename of the h5m file created
+    """
+    output_filenames = [
+        h5m_filename,
+        trelis_filename,
+        cubit_filename,
+        geometry_details_filename]
+    filenames_extensions = ['.h5m', '.trelis', '.cub', '.json']
+
+    path_output_filenames = []
+
+    for output_file, extension in zip(output_filenames, filenames_extensions):
+
+        if output_file is not None:
+            path_filename = Path(output_file)
+
+            if path_filename.suffix != extension:
+                path_filename = path_filename.with_suffix(extension)
+
+            path_filename.parents[0].mkdir(parents=True, exist_ok=True)
+
+            path_output_filenames.append(str(path_filename))
+
+    shutil.copy(
+        src=Path(__file__).parent.absolute() / Path('parametric_neutronics') /
+        'make_faceteted_neutronics_model.py',
+        dst=Path().absolute()
+    )
+
+    if not Path("make_faceteted_neutronics_model.py").is_file():
+        raise FileNotFoundError(
+            "The make_faceteted_neutronics_model.py was not found in the \
+            directory")
+
+    os.system('rm dagmc_not_watertight.h5m')
+
+    if batch:
+        trelis_cmd = 'trelis -batch -nographics'
+    else:
+        trelis_cmd = 'trelis'
+
+    os.system(
+        trelis_cmd +
+        " make_faceteted_neutronics_model.py \"faceting_tolerance='" +
+        str(faceting_tolerance) +
+        "'\" \"merge_tolerance='" +
+        str(merge_tolerance) +
+        "'\" \"material_key_name='" +
+        str(material_key_name) +
+        "'\" \"geometry_key_name='" +
+        str(geometry_key_name) +
+        "'\" \"h5m_filename='" +
+        str(h5m_filename) +
+        "'\" \"manifest_filename='" +
+        str(manifest_filename) +
+        "'\" \"cubit_filename='" +
+        str(cubit_filename) +
+        "'\" \"trelis_filename='" +
+        str(trelis_filename) +
+        "'\" \"geometry_details_filename='" +
+        str(geometry_details_filename) +
+        "'\" \"surface_reflectivity_name='" +
+        str(surface_reflectivity_name) +
+        "'\"")
+
+    os.system('rm make_faceteted_neutronics_model.py')
+
+    if not Path(h5m_filename).is_file():
+        raise FileNotFoundError(
+            "The h5m file " + h5m_filename + " was not found \
+            in the directory, the Trelis stage has failed")
+
+    return path_output_filenames
+
+
+def make_watertight(
+        input_filename: str = "dagmc_not_watertight.h5m",
+        output_filename: str = "dagmc.h5m",
+) -> str:
+    """Runs the DAGMC make_watertight executable that seals the facetets of
+    the geometry with specified input and output h5m files. Not needed for
+    h5m file produced with pymoab method.
+
+    Arguments:
+        input_filename: the non watertight h5m file to make watertight.
+        output_filename: the filename of the watertight h5m file.
+
+    Returns:
+        The filename of the h5m file created
+    """
+
+    if not Path(input_filename).is_file():
+        raise FileNotFoundError("Failed to find {}".format(input_filename))
+
+    os.system('rm {}'.format(output_filename))
+
+    try:
+        output = subprocess.check_output(
+            "make_watertight {} -o {}".format(input_filename, output_filename),
+            shell=True,
+            universal_newlines=True,
+        )
+        print(output)
+    except BaseException:
+        raise NameError(
+            "make_watertight failed, check DAGMC is install and the DAGMC/bin "
+            "folder is in the path directory (Linux and Mac) or set as an "
+            "enviromental varible (Windows)")
+
+    if not Path(output_filename).is_file():
+        raise FileNotFoundError("Failed to produce dagmc.h5m")
+
+    return output_filename
+
+
+def export_vtk(
+    h5m_filename: str,
+    filename: Optional[str] = 'dagmc.vtk',
+    include_graveyard: Optional[bool] = False
+):
+    """Produces a vtk geometry compatable from the dagmc h5m file. This is
+    useful for checking the geometry that is used for transport.
+
+    Arguments:
+        filename: filename of vtk outputfile. If the filename does not end
+            with .vtk then .vtk will be added.
+        h5m_filename: filename of h5m outputfile. If the filename does not
+            end with .h5m then .h5m will be added.
+        include_graveyard: optionally include the graveyard in the vtk file
+
+    Returns:
+        filename of the vtk file produced
+    """
+
+    path_h5m_filename = Path(h5m_filename)
+    if path_h5m_filename.suffix != ".h5m":
+        path_h5m_filename = path_h5m_filename.with_suffix(".h5m")
+
+    if path_h5m_filename.is_file() is False:
+        raise FileNotFoundError(
+            'h5m_filename not found in location', path_h5m_filename
+        )
+
+    path_filename = Path(filename)
+    if path_filename.suffix != ".vtk":
+        path_filename = path_filename.with_suffix(".vtk")
+
+    if include_graveyard:
+        tags_to_remove = None
+    else:
+        tags_to_remove = [
+            'graveyard.stp',
+            'mat:graveyard',
+            'reflective',
+            'mat:vacuum',
+            'cuttingwedge.stp']
+    remove_tags(
+        input=str(path_h5m_filename),
+        output=str(path_filename),
+        tags=tags_to_remove
+    )
+
+    return str(path_filename)
+
+
+def define_moab_core_and_tags():
+    """Creates a MOAB Core instance which can be built up by adding sets of
+    triangles to the instance
+
+    Returns:
+        (pymoab Core): A pymoab.core.Core() instance
+        (pymoab tag_handle): A pymoab.core.tag_get_handle() instance
+    """
+
+    try:
+        from pymoab import core, types
+    except ImportError:
+        raise ImportError(
+            'PyMoab not found, export_h5m method is not available')
+
+    # create pymoab instance
+    moab_core = core.Core()
+
+    tags = dict()
+
+    sense_tag_name = "GEOM_SENSE_2"
+    sense_tag_size = 2
+    tags['surf_sense'] = moab_core.tag_get_handle(
+        sense_tag_name,
+        sense_tag_size,
+        types.MB_TYPE_HANDLE,
+        types.MB_TAG_SPARSE,
+        create_if_missing=True)
+
+    tags['category'] = moab_core.tag_get_handle(
+        types.CATEGORY_TAG_NAME,
+        types.CATEGORY_TAG_SIZE,
+        types.MB_TYPE_OPAQUE,
+        types.MB_TAG_SPARSE,
+        create_if_missing=True)
+    tags['name'] = moab_core.tag_get_handle(
+        types.NAME_TAG_NAME,
+        types.NAME_TAG_SIZE,
+        types.MB_TYPE_OPAQUE,
+        types.MB_TAG_SPARSE,
+        create_if_missing=True)
+    tags['geom_dimension'] = moab_core.tag_get_handle(
+        types.GEOM_DIMENSION_TAG_NAME,
+        1,
+        types.MB_TYPE_INTEGER,
+        types.MB_TAG_DENSE,
+        create_if_missing=True)
+
+    # Global ID is a default tag, just need the name to retrieve
+    tags['global_id'] = moab_core.tag_get_handle(types.GLOBAL_ID_TAG_NAME)
+
+    return moab_core, tags
+
+
+def add_stl_to_moab_core(
+        moab_core,
+        surface_id: int,
+        volume_id: int,
+        material_name: str,
+        tags,
+        stl_filename: str):
+    """Computes the m and c coefficients of the equation (y=mx+c) for
+    a straight line from two points.
+
+    Args:
+        moab_core (pymoab.core.Core):
+        surface_id (int): the id number to apply to the surface
+        volume_id (int): the id numbers to apply to the volumes
+        material_name (str): the material tag name to add. the value provided
+            will be prepended with "mat:" unless it is "reflective" which is
+            a special case and therefore will remain as is.
+        tags (pymoab tag_handle): the MOAB tags
+        stl_filename (str): the filename of the stl file to load into the moab
+            core
+
+    Returns:
+        (pymoab Core): An updated pymoab.core.Core() instance
+    """
+
+    surface_set = moab_core.create_meshset()
+    volume_set = moab_core.create_meshset()
+
+    # recent versions of MOAB handle this automatically
+    # but best to go ahead and do it manually
+    moab_core.tag_set_data(tags['global_id'], volume_set, volume_id)
+
+    moab_core.tag_set_data(tags['global_id'], surface_set, surface_id)
+
+    # set geom IDs
+    moab_core.tag_set_data(tags['geom_dimension'], volume_set, 3)
+    moab_core.tag_set_data(tags['geom_dimension'], surface_set, 2)
+
+    # set category tag values
+    moab_core.tag_set_data(tags['category'], volume_set, "Volume")
+    moab_core.tag_set_data(tags['category'], surface_set, "Surface")
+
+    # establish parent-child relationship
+    moab_core.add_parent_child(volume_set, surface_set)
+
+    # set surface sense
+    sense_data = [volume_set, np.uint64(0)]
+    moab_core.tag_set_data(tags['surf_sense'], surface_set, sense_data)
+
+    # load the stl triangles/vertices into the surface set
+    moab_core.load_file(stl_filename, surface_set)
+
+    group_set = moab_core.create_meshset()
+    moab_core.tag_set_data(tags['category'], group_set, "Group")
+
+    # reflective is a special case that should not have mat: in front
+    if not material_name == 'reflective':
+        dag_material_tag = "mat:{}".format(material_name)
+    else:
+        dag_material_tag = material_name
+
+    moab_core.tag_set_data(
+        tags['name'],
+        group_set,
+        dag_material_tag
+    )
+    moab_core.tag_set_data(tags['geom_dimension'], group_set, 4)
+
+    # add the volume to this group set
+    moab_core.add_entity(group_set, volume_set)
+
+    return moab_core
 
 
 def transform_curve(edge, tolerance: float = 1e-3):
